💄 move tao dashboard network switch next to period tabs
